feat: Add API routes for account-related operations
This update implements APIs and the initial logic for working with
accounts. It currently handles Sub Accounts, Family Accounts, Fund
Requests and another API for retrieving both sub accounts and family
accounts in one response. It is fixing issue 'INT-244'.

This is not the final yet as there are some dependencies I am waiting
for.